use Asciidoc for termdef boilerplate: https://github.com/metanorma/metanorma-iec/issues/339
