+++ lib/metanorma/standoc/cleanup_boilerplate.rb
@@ -1,41 +1,8 @@
+require_relative "cleanup_terms_boilerplate"
+
 module Metanorma
   module Standoc
     module Cleanup
-      def external_terms_boilerplate(sources)
-        e = @i18n.external_terms_boilerplate
-        @i18n.l10n(e.gsub(/%(?=\p{P}|\p{Z}|$)/, sources || "???"),
-                   @lang, @script, @locale)
-      end
-
-      def internal_external_terms_boilerplate(sources)
-        e = @i18n.internal_external_terms_boilerplate
-        @i18n.l10n(e.gsub(/%(?=\p{P}|\p{Z}|$)/, sources || "??"),
-                   @lang, @script)
-      end
-
-      def term_defs_boilerplate(div, source, term, _preface, isodoc)
-        a = @i18n.term_def_boilerplate and div.next = a
-        source.each do |s|
-          @anchors[s["bibitemid"]] or
-            @log.add("Crossreferences", nil,
-                     "term source #{s['bibitemid']} not referenced", severity: 1)
-        end
-        a = if source.empty? && term.nil? then @i18n.no_terms_boilerplate
-            else term_defs_boilerplate_cont(source, term, isodoc)
-            end and div.next = a
-      end
-
-      def term_defs_boilerplate_cont(src, term, isodoc)
-        sources = isodoc.sentence_join(src.map do |s|
-          %{<eref bibitemid="#{s['bibitemid']}"/>}
-        end)
-        if src.empty? then @i18n.internal_terms_boilerplate
-        elsif term.nil? then external_terms_boilerplate(sources)
-        else
-          internal_external_terms_boilerplate(sources)
-        end
-      end
-
       def norm_ref_preface(ref)
         ins = norm_ref_boilerplate_insert_location(ref)
         ins2 = norm_ref_process_boilerplate_note(ref)
@@ -69,30 +36,21 @@ module Metanorma
         ref.at("./title")
       end
 
-      TERM_CLAUSE =
-        "//sections//terms[not(.//ancestor::clause[@type = 'terms'])] | " \
-        "//sections/clause[descendant::terms][@type = 'terms'] | " \
-        "//sections/clause[not(@type = 'terms')]//terms".freeze
-
       NORM_REF =
         "//bibliography/references[@normative = 'true'][not(@hidden)] | " \
         "//bibliography/clause[.//references[@normative = 'true']]".freeze
 
       def boilerplate_isodoc(xmldoc)
+        # prevent infinite recursion of asciidoc boilerplate processing
+        # in termdef_boilerplate_insert and initial_boilerplate
+        xmldoc.at("//metanorma-extension/semantic-metadata/" \
+                  "headless[text() = 'true']") and return nil
         x = xmldoc.dup
         x.root.add_namespace(nil, self.class::XML_NAMESPACE)
         xml = Nokogiri::XML(x.to_xml)
         @isodoc ||= isodoc(@lang, @script, @locale)
         @isodoc.info(xml, nil)
         @isodoc
-      end
-
-      def termdef_boilerplate_cleanup(xmldoc)
-        # termdef_remove_initial_paras(xmldoc)
-      end
-
-      def termdef_remove_initial_paras(xmldoc)
-        xmldoc.xpath("//terms/p | //terms/ul").each(&:remove)
       end
 
       def unwrap_boilerplate_clauses(xmldoc, xpath)
@@ -105,51 +63,9 @@ module Metanorma
         end
       end
 
-      def termdef_boilerplate_insert(xmldoc, isodoc, once = false)
-        if once
-          f = termdef_boilerplate_insert_location(xmldoc) and
-            termdef_boilerplate_insert1(f, xmldoc, isodoc)
-        else
-          xmldoc.xpath(self.class::TERM_CLAUSE).each do |f|
-            termdef_boilerplate_insert1(f, xmldoc, isodoc)
-          end
-        end
-      end
-
-      def termdef_boilerplate_insert_location(xmldoc)
-        f = xmldoc.at(self.class::TERM_CLAUSE)
-        root = xmldoc.at("//sections/terms | //sections/clause[@type = 'terms']")
-        if f && root && f["id"] != root["id"]
-          f = termdef_boilerplate_climb_up(f, root)
-        elsif !f && root then f = root
-        end
-        f
-      end
-
-      def termdef_boilerplate_climb_up(clause, container)
-        container.at(".//*[@id = '#{clause['id']}']") or return clause
-        while (n = clause.parent)
-          n.at(".//definitions") and break
-          clause = n
-          n["id"] == container["id"] and break
-        end
-        clause
-      end
-
-      def termdef_boilerplate_insert1(sect, xmldoc, isodoc)
-        ins = sect.at("./title")
-        if (ins2 = sect.at("./clause[@type = 'boilerplate'] | " \
-                "./note[@type = 'boilerplate']"))
-          ins2.text.strip.downcase == "(default)" or return
-          ins2.children = " "
-          ins = ins2.children.first
-        end
-        term_defs_boilerplate(ins, xmldoc.xpath(".//termdocsource"),
-                              sect.at(".//term"), sect.at(".//p"), isodoc)
-      end
-
       def boilerplate_cleanup(xmldoc)
-        isodoc = boilerplate_isodoc(xmldoc)
+        isodoc = boilerplate_isodoc(xmldoc) or return
+        term_children_cleanup(xmldoc)
         termdef_boilerplate_cleanup(xmldoc)
         termdef_boilerplate_insert(xmldoc, isodoc)
         unwrap_boilerplate_clauses(xmldoc, self.class::TERM_CLAUSE)
@@ -245,6 +161,7 @@ module Metanorma
           /^_\d+$/.match?(n["id"]) and
             n["id"] = "_#{UUIDTools::UUID.random_create}"
         end
+        xml
       end
 
       def boilerplate_top_elements(xml)
