+++ lib/metanorma/standoc/cleanup_terms.rb
@@ -121,7 +121,7 @@ module Metanorma
         termdefinition_cleanup(xmldoc)
         termdomain1_cleanup(xmldoc)
         termnote_example_cleanup(xmldoc)
-        term_children_cleanup(xmldoc)
+        #term_children_cleanup(xmldoc)
       end
 
       def term_index_cleanup(xmldoc)
